refactor: installStateLayer
