+++ packages/web-widget/src/state-layer.ts
@@ -1,3 +1,9 @@
+import {
+  callContext,
+  createContext,
+  useAllWidgetState,
+} from "@web-widget/schema/client-helpers";
+
 export type StateLayerHandler = (
   value: any,
   index: number,
@@ -17,5 +23,32 @@ export class StateLayer implements ArrayLike<any> {
     const length = Array.prototype.push.apply(this, list);
     list.forEach(this.#handler);
     return length;
+  }
+}
+
+export function installStateLayer(callback: () => void) {
+  const stateElement = document.querySelector(
+    `script[name="state\\:web-router"]`
+  ) as HTMLScriptElement;
+  const context = stateElement
+    ? JSON.parse(stateElement.textContent as string)
+    : {};
+
+  callContext(createContext(context), () => {
+    const currentState = self.stateLayer as unknown as undefined | any[];
+    const allState = useAllWidgetState();
+    self.stateLayer = new StateLayer((item) => Object.assign(allState, item));
+
+    if (currentState) {
+      self.stateLayer.push(...currentState);
+    }
+
+    callback();
+  });
+}
+
+declare global {
+  interface Window {
+    stateLayer: StateLayer;
   }
 }
